+++ backend/keystone.ts
@@ -10,7 +10,10 @@ import 'dotenv/config';
 import { config } from '@keystone-6/core';
 
 // to keep this file tidy, we define our schema in a different file
-import { lists } from './schema';
+import {
+  lists,
+  extendGraphqlSchema,
+} from './schema';
 
 // authentication is configured separately here too, but you might move this elsewhere
 // when you write your list-level access control functions, as they typically rely on session data
@@ -50,5 +53,6 @@ export default withAuth(
     },
     lists,
     session,
+    extendGraphqlSchema,
   })
 );