+++ frontend/src/app/course/[id]/results/[testId]/page.js
@@ -1,0 +1,15 @@
+
+export default function TestIdPage({
+  params,
+}) {
+
+  const {
+    testId,
+  } = params;
+
+  return (
+    <main>
+      <p>Go get dis test info: {testId}</p>
+    </main>
+  )
+}