+++ frontend/src/utils/useForm.js
@@ -22,6 +22,12 @@ export default function useForm(initial = {}) {
       value = parseInt(value);
     }
 
+    console.log({
+      value,
+      name,
+      type,
+    });
+
     setInputs({
       ...inputs,
       [name]: value,