+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] - 2024-03-14
+### Added
+- Frontend
+  - Dashboard has previous test results, available courses and link to profile
+  - Can take course and submit answers
+- Backend
+  - Starting course creates new test
+  - Can submit questions and answers to get results
+
+
 ## [0.1.1] - 2024-03-13
 ### Added
 - Frontend